+++ crates/render-lib/src/benching/rendering.rs
@@ -14,7 +14,7 @@ pub fn write_frametimes_svgs<P>(
     renderer: &mut dyn Renderer,
     svg_dir: P,
     output: P,
-    frames: usize,
+    frames: u64,
 ) -> Result<()>
 where
     P: Into<PathBuf>,
@@ -63,7 +63,7 @@ pub fn write_frametimes_primitives<P>(
     primitives: &Vec<(String, Primitive)>,
     count: u32,
     output: P,
-    frames: usize,
+    frames: u64,
 ) -> Result<()>
 where
     P: Into<PathBuf>,
